Use commonjs instead of module
