+++ tsconfig.json
@@ -21,5 +21,5 @@
         "noUnusedLocals": true,
         "noUnusedParameters": true
     },
-    "exclude": ["./scripts/create-token.mjs", "./examples/"]
+    "exclude": ["./scripts/create-token.js", "./examples/"]
 }
